rename: types/authentication.ts -> auth.ts
