+++ client/src/components/meetups/MeetupFilterChoices/index.tsx
@@ -1,5 +1,5 @@
 import styles from './MeetupFilterChoices.module.scss';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useSetRecoilState } from 'recoil';
 import { meetupBoardFiltersState, meetupFiltersState } from 'states/meetup';
 import OptionChip from 'components/common/OptionChip';
 import { FilterOptionData } from 'types/common';
